refactor: LottoGame 클래스 분리
